refactor(FolderNameList): Refactor FolderNameList
