+++ src/components/FolderNameList/FolderNameList.module.css
@@ -1,5 +1,5 @@
 .folder-buttons-section {
-  padding-right: 20rem;
+  padding: 2rem 15rem 2rem 4rem;
 }
 
 .folder-buttons-list {
